Major state and loss resetting improvement
Suppose there are two videos, "a" and "b", so far we were trying to estimate:

   1) a[-1] ~= net(a[-2])
   2) b[ 0] ~= net(a[-1]) <-- this has to be removed
   3) b[ 1] ~= net(b[ 0])

Therefore, we now force "net(a[-1]) = b[0]", so there is no error or gradient
associated to it.
Moreover, we do not want a gradient from 2) to 1) through the state forwarding,
so we kill the state 1) -> 2) and therefore the gradient 2) -> 1).
Furthermore, we reset the state 2) -> 3) because the state refers to "a",
and not to "b".
Finally, we force also "net(a[0]) = a[1]" and "net(b[0]) = b[1]", given that
there was no state or the state has been just reset.